+++ NEWS.md
@@ -16,6 +16,9 @@
 * `safe_digest()` uses a better strategy, and returns NA for directories 
   (#138, #146).
 
+* `colourise` now uses option `testthat.use_colours` (default: `TRUE`). If it
+  is `FALSE`, output is not colourised.
+
 * Random praise is renabled by default (again!) (#164).
 
 * `is_identical_to()` only calls `all.equal()` to generate an informative